layered: Movd model order comparator in separate classes #608
